pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-gpt5-bot"
version = "0.1.0"
description = "Discord bot (text + image input -> text output)"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "toml>=0.10.2",
  "discord.py>=2.3.2",
  "openai>=1.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.4.8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C90"]
ignore = []

